feat(backend): basic implementation
This is what we will work with troughout the semester. There is a lot of functionalities that can be added here, such as authentication with Google, mailing etc. For now, this is enough to have a basic understanding of the tech stack that will be used in this part of the project. There are also some things that need improvement, like the way the user can sign-in (there are some security issues there) or the documentation. I ask to read the code patiently, and ask for assistance if needed.